Bring back QR Code class and update base code
